requirements-dev.txt
-r requirements.txt
pytest==9.0.0
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-homeassistant-custom-component==0.13.325
aioresponses>=0.7.6
syrupy>=5.0.0
ruff>=0.15.0
mypy>=1.15.0
mutmut>=3.0.0
pip-audit>=2.7.0
types-pyyaml>=6.0.0
pip-tools>=7.0.0
